Measure chart callbacks in unit coverage
